+++ spec/directions/bake_chapter_glossary/v1_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Kitchen::Directions::BakeChapterGlossary::V1 do
   before do
     stub_locales({
       'eoc_key_terms_title': 'Key Terms',
-      'eoc_exercises_title': 'Exercises'
+      'eoc_composite_metadata_title': 'Chapter Review'
     })
   end
 
@@ -112,7 +112,7 @@ RSpec.describe Kitchen::Directions::BakeChapterGlossary::V1 do
                 <span class="os-text">Key Terms</span>
               </h3>
               <div data-type="metadata" style="display: none;">
-                <h1 data-type="document-title" itemprop="name">Exercises</h1>
+                <h1 data-type="document-title" itemprop="name">Chapter Review</h1>
                 <div class="authors" id="authors_copy_1">Authors</div><div class="publishers" id="publishers_copy_1">Publishers</div><div class="print-style" id="print-style_copy_1">Print Style</div><div class="permissions" id="permissions_copy_1">Permissions</div><div data-type="subject" id="subject_copy_1">Subject</div>
               </div>
               <dl>